requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
matplotlib>=3.8
PyYAML>=6.0
joblib>=1.3
streamlit>=1.33
streamlit>=1.33
